pyproject.toml
[project]
name = "meeting-police"
version = "0.1.0"
description = "Meeting facilitation platform backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "boto3",
    "python-dotenv",
    "httpx",
    "PyJWT",
    "cryptography",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
